requirements.txt
streamlit
pdfplumber
requests
google-genai
